move some testing helpers into the common module
